+++ examples/neotrellis_ripples.rs
@@ -2,7 +2,8 @@
 #![no_main]
 #![allow(incomplete_features)]
 #![feature(generic_const_exprs)]
-/// https://github.com/adafruit/Adafruit_Seesaw/blob/master/examples/NeoTrellis/ripples/ripples.ino
+/// Arduino example: https://github.com/adafruit/Adafruit_Seesaw/blob/master/examples/NeoTrellis/ripples/ripples.ino
+/// Demo video: https://storage.googleapis.com/apemedia/neotrellis576.mp4
 use adafruit_seesaw::{devices::NeoTrellis, prelude::*, SeesawRefCell};
 use core::ops::{Add, AddAssign, Mul};
 use cortex_m_rt::entry;